If files already exist, skip download and extraction respectively

+++ scraping/0_extract_ars.py
@@ -14,30 +14,38 @@ import requests
 import pandas as pd
 from pathlib import Path
 
-def download_excel(url: str, save_dir: str, filename: str) -> str:
+def download_file(url: str, save_dir: str, filename: str) -> str:
     """
-    Download an excel from the specified URL and save it to the specified directory.
+    Download a file from the URL and save it to the specified directory.
 
     :param url: URL of the file to download.
     :param save_dir: Directory where the downloaded file will be saved.
     :param file_name: Name of the file to save.
     :return: If the download was sucessfull path to file, else None.
     """
+    # Create the save directory if it does not yet exist
     Path(save_dir).mkdir(parents=True, exist_ok=True)
+
     save_path = os.path.join(save_dir, filename)
+
+    # If file already exists locally, skip the download
+    if os.path.isfile(save_path):
+        print(f"File already exists at {save_path}. Skipping download.")
+        return save_path
     
+    # Download file from the specified URL
     try:
         response = requests.get(url)
         response.raise_for_status()
         with open(save_path, 'wb') as file:
             file.write(response.content)
-        print(f"Excel file downloaded and saved to {save_path}")
+        print(f"File downloaded and saved to {save_path}.")
         return save_path
     except requests.RequestException as e:
         print(f"Failed to download file: {e}")
         return None
 
-def extract_ars(excel_path: str, save_dir: str, filename: str):
+def extract_ars(excel_path: str, save_dir: str, filename: str) -> None:
     """
     Extract the ARS for all German municipalities from the excel at the 
     specified location and save as csv.
@@ -47,25 +55,34 @@ def extract_ars(excel_path: str, save_dir: str, filename: str):
     :param filename: name of the file to save.
     :return: None
     """
+    # Create the save directory if it does not yet exist
     Path(save_dir).mkdir(parents=True, exist_ok=True)
     save_path = os.path.join(save_dir, filename)
 
+    # If file already exists, skip the extraction
+    if os.path.isfile(save_path):
+        print(f"File already exists at {save_path}. Skipping extraction.")
+        return None
+
+    # Extract ARS from excel file and save to CSV
     try:
         df = pd.read_excel(excel_path)
-        column_data = df.iloc[8:, 1].dropna()  
+        column_data = df.iloc[7:, 1].dropna()  
         column_data.to_csv(save_path, index=False, header=False)
         print(f"CSV with ARS saved to: {save_path}")
     except Exception as e:
         print(f"Failed to extract ARS: {e}")
 
-if __name__ == "__main__":
-
-    EXCEL_URL = "https://www.xrepository.de/api/xrepository/urn:de:bund:destatis:bevoelkerungsstatistik:schluessel:rs_2024-10-31/download/Regionalschl_ssel_2024-10-31.xlsx"
+def main():
+    EXCEL_URL = "https://www.xrepository.de/api/xrepository/urn:de:bund:destatis:bevoelkerungsstatistik:schluessel:rs_2024-10-31/download/Regionalschl_ssel_2024-10-31.xlsx"  
     
-    save_dir = "data/ars"
+    save_dir = "scraping/data/ars"
     
-    excel_file_path = download_excel(EXCEL_URL, save_dir, filename="ars_raw.xlsx")
+    filepath = download_file(EXCEL_URL, save_dir, filename="ars_raw.xlsx")
     
     # If the download was successful, extract the ARS
-    if excel_file_path:
-        extract_ars(excel_file_path, save_dir, filename="ars_only.csv")
+    if filepath:
+        extract_ars(filepath, save_dir, filename="ars_only.csv")
+
+if __name__ == "__main__":
+    main()
